Implemented game logic, used mixin to keep gamescreen alive in case of page switching, made play button reset the state
Implemented game logic: clicking on the continue button in an active row of pins now checks the input code against the code created when the game starts, evaluates and updates the scorepins, makes the row inactive and creates a new row.

Game screen state: implemented a mixin to keep the state of the game alive when switching to info/settings screen, therefore not losing any inputs and keeping the rows that have been played inactive.

Start screen button now resets the value of allrows, in effect starting a new game when clicked on whilst in an active game